requirements.txt
backcall>=0.1.0
biopython>=1.72
cbmpy>=0.7.25
certifi>=2018.10.15
colorama>=0.4.0
decorator>=4.3.0
ipython>=7.1.1
ipython-genutils>=0.2.0
jedi>=0.13.1
numpy>=1.15.4
nose>=1.3.7
parso>=0.4.0
pickleshare>=0.7.5
prompt-toolkit>=2.0.7
Pygments>=2.2.0
pyparsing>=2.3.0
PyQt5>=5.11.3
PyQt5-sip>=4.19.13
python-libsbml>=5.17.0
sip>=4.19.8
six>=1.11.0
traitlets>=4.3.2
wcwidth>=0.1.7
wincertstore>=0.2
xlrd>=1.1.0
xlwt>=1.3.0
